Add files via upload
This project  serve as an introduction to data analysis with python. In this note book, i have used Python programming to import data, clean the data and  to conduct exploratory analysis using boxplot and regplot. I went further and developed the model using  Linear Regression. I used Pipelines to predict the price of the houses . As for model evaluation and refinement i used cross validation and Ridge Regression..